refactor request binders
Refactor the JSON request binder so it doesn't change the API or
introduce new exported variables, especially since the previous
implementation didn't make it clear that the Json variable was intended
for internal use only in order to facilitate testing. This change also
makes it so the documentation doesn't have to change.

Make the RequestBinder unexported (RequestBinder -> requestBinder) since it's not
intended to be used by API clients.

Create requestBinders for form-urlencoded and multipart/form-data.