Don't convert non-binding.Errors.
As a follow-on to 52f4dd9, continue the process of allowing errors that
aren't binding.Errors to flow through without adaptation.

The need for this arises from cases where validation is performed
against a struct that contains field of a struct type that also
implements Validator, when the validation for the inner type returns an
error that isn't a binding.Errors. The existing code consumes such
errors and turns them into not-particularly-helpful binding.Errors
instances, inhibiting upstack processing of distinct error types. This
does away with that interference.